Remove AIEditorManager
fixes #14873

- move the customizations to the parent class EditorPreviewManager
- rename AIEditorSelectionResolver and move it to editor-preview-manager.ts
