+++ packages/editor-preview/src/browser/editor-preview-frontend-module.ts
@@ -18,11 +18,11 @@ import '../../src/browser/style/index.css';
 import { FrontendApplicationContribution, KeybindingContribution, WidgetFactory } from '@theia/core/lib/browser';
 import { ContainerModule } from '@theia/core/shared/inversify';
 import { bindEditorPreviewPreferences } from './editor-preview-preferences';
-import { EditorPreviewManager } from './editor-preview-manager';
+import { EditorPreviewManager, EditorSelectionResolver } from './editor-preview-manager';
 import { EditorManager } from '@theia/editor/lib/browser';
 import { EditorPreviewWidgetFactory } from './editor-preview-widget-factory';
 import { EditorPreviewContribution } from './editor-preview-contribution';
-import { CommandContribution, MenuContribution } from '@theia/core/lib/common';
+import { CommandContribution, MenuContribution, bindContributionProvider } from '@theia/core/lib/common';
 import { OpenEditorsTreeDecorator } from '@theia/navigator/lib/browser/open-editors-widget/navigator-open-editors-decorator-service';
 import { EditorPreviewTreeDecorator } from './editor-preview-tree-decorator';
 
@@ -43,4 +43,5 @@ export default new ContainerModule((bind, unbind, isBound, rebind) => {
     bind(OpenEditorsTreeDecorator).toService(EditorPreviewTreeDecorator);
     bind(FrontendApplicationContribution).toService(EditorPreviewTreeDecorator);
     bindEditorPreviewPreferences(bind);
+    bindContributionProvider(bind, EditorSelectionResolver);
 });
